refactor: move lib/gui/Diagnostic => apps/deskflow-gui/Diagnostic
