+++ src/apps/deskflow-gui/Diagnostic.cpp
