+++ src/apps/deskflow-gui/Diagnostic.h
